Make it easier to inject alternate providers.
When using services that have not been explicitly named, if there is more than class or service that is defined for the FakerProvider
interface, the DI resolver finds all of those services when trying to find the FakerProvider to inject into the ProviderExtension
constructor. This situation includes when the services configuration has a service that uses a factory to create an instance of an
interface (where the interface is given as the 'class' parameter for the service) as well as when it has a service for a completely
different class, that just happens to implement that interface.

This change attempts to alleviate this issue. On one side, we explicitly name all classes and services that can provide a Faker
Generator. On the other side, where we inject the provider to the ProviderExtension, we explicitly name a service to use for that
provider argument. If we want to change the provider to inject, we change the argument.

This is still slightly messy, however, as it would require redefining the entire service. Instead, we have one more service defined.
We define a 'pointer' service that names the actual provider service that we wish to use and we instead inject that 'pointer' service
into the extension. This way, we only have to change the definition of that 'pointer' service, instead of fully redefining the
ProviderExtension.

It is possible that we could have also explicitly named the FakerProviderFactory etc. but for now, at least, this solves the problem.