FakerProvider that will pull the Generator from a defined PSR-11 compatible DI container file.
As an example... if using Symfony, defining:
parameters:
  fakerstan:
    psr:
      phpContainerPath: ./var/cache/dev/App_KernelDevDebugContainer.php
      returnsContainer: true

tells Fakerstan to load the Symfony DI container from the dev environment, and that that container
file will return the container object. The default containerFakerId setting is correct for the ID
of the Generator in the Symfony container, and so Fakerstan will load the Generator successfully.

By using the Generator from the DI container we ensure that all custom providers are added to the
Generator first and so making them available for analysis.