FakerProvider that pulls the Generator from a PSR-11 container.
As it stands, this code only supports pulling the container itself from a PHP file. That file may
either return the container object directly, or set it into a defined variable name. The provider
will include the PHP file defined to get the container, then return the Generator from the container
by using the defined ID.

By using the Generator from the DI container we ensure that all custom providers are added to the
Generator first and so making them available for analysis.

As an example... if using Symfony, something similar to:
    $provider = new PsrContainerFakerProvider(
        './var/cache/dev/App_KernelDevDebugContainer.php',
        null,
        'Faker\Generator'
    );

will give us a provider that, when asked for the Faker Generator, will include the
App_KernelDevDebugContainer.php file, using the return value as the container. From that container
the provider will then pull a service with the ID 'Faker\Generator' and return it.