A factory that can provide a PsrContainerFakerProvider.
To allow for the service that provides \CalebDW\Fakerstan\FakerProvider to call the
static 'create' method on the defined factory, this factory takes parameters for its
constructor and then copies those values onto static member variables. The static
'create' method then has access to the values that were given to the constructor.